% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{docExampleData}
\alias{docExampleData}
\title{Documentation Example Data}
\format{
## `docExampleData`
A csv with 6 observations on the following 3 variables.
\describe{
  \item{FishID}{a numeric vector}
  \item{Length}{a numeric vector}
  \item{Age}{a numeric vector}
  ...
}
}
\source{
generateDataset function
}
\usage{
docExampleData
}
\description{
Fake data used to verify examples in function documentation
Report ...
}
\keyword{datasets}
