% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMGroup1-package.R
\docType{package}
\name{EMGroup1-package}
\alias{EMGroup1}
\alias{EMGroup1-package}
\title{EMGroup1: Computes EM Algorithm for FishLengths Data}
\description{
Performs the Expectation-Maximisation (EM) Algotithm on Fish Lengths Data that has age groups assigned to only a portion of the data. Function teamEM performs all steps of the optimisation in one call, returning the optimised mu, sigma, and labmda parameters for Age groups 1, 2, and 3.
}
\author{
\strong{Maintainer}: Narayan Murti \email{nam27@st-andrews.ac.uk}

Authors:
\itemize{
  \item Lee Scott \email{ls318@st-andrews.ac.uk}
  \item Ivor Walker \email{iw72@st-andrews.ac.uk}
  \item Yi Zhang \email{yz209@st-andrews.ac.uk}
}

}
\keyword{internal}
