% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGenerator.R
\name{generateDataset}
\alias{generateDataset}
\title{Generate Random Dataset for Testing}
\usage{
generateDataset(seed, fish, datasize, inits, distributions)
}
\arguments{
\item{seed}{Set the random seed.}

\item{fish}{Specify the amount of fish.}

\item{datasize}{Specify the amount of length observations. The size of the dataset.}

\item{inits}{Initialisation parameters from FishLengths dataset.}

\item{distributions}{A length-4 vector with probabilities for the occurance of (NA, Age1, Age2, and Age3), respectively.}
}
\value{
A simulated dataset with columns FishID, Length, and Age
}
\description{
Generate Random Dataset for Testing
}
\examples{
known <- docExampleData[!is.na(docExampleData$Age), ] # known data is not NA
unknown <- docExampleData[is.na(docExampleData$Age), ] # unknown data is NA
sorted_data <- rbind(known, unknown) # combine data
inits <- initialise(known, unknown, sorted_data)

generateDataset(0, 100, 1000, inits, c(0.9, 0.04, 0.03, 0.03))
}
