% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x}
\alias{x}
\title{Fish Lengths Data}
\format{
## `FishLengths`
A data frame with 1000 observations on the following 3 variables.
\describe{
  \item{FishID}{a numeric vector}
  \item{Length}{a numeric vector}
  \item{Age}{a numeric vector}
  ...
}
}
\source{
<https://moody.st-andrews.ac.uk/moodle/mod/resource/view.php?id=1267766>
}
\usage{
x
}
\description{
Data of fish of different lengths and their associated age group. 100 observations contain an age group assignments, 900 observations contain an unassigned age group.
Report ...
}
\keyword{datasets}
