% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{expector}
\alias{expector}
\title{Expectation Step for EM Algorithm}
\usage{
expector(known, sorted_data, estimates)
}
\arguments{
\item{known}{Data frame with columns for Length and Age, only for known Age}

\item{sorted_data}{Data frame with columns for Length and Age, with all observations with known Age at the top, and all unknown Age obervations at the bottom.}

\item{estimates}{3x3 matrix / dataframe with row 1 containing mu estimates, row 2 containing sd   estimates, and row 3 containg lambda estimates all for age group j}
}
\value{
A list containing posteriors and densities
\itemize{
 \item posteriors - Data frame containing posterior probabilities that a length belongs to each age group, for each length in the data
 \item densities - Gaussian probability densities of each length based on the current estimated parameters
}
}
\description{
Expectation Step for EM Algorithm
}
\examples{
data <- read_csv("data/docExampleData.csv")
estimates <- data.frame(matrix(c(5, 10, 15, 5, 6, 7, .5, .6, .7),
                              nrow = 3,
                              ncol = 3))
colnames(estimates) <- c("mu", "sigma", "lambda")
rownames(estimates) <- c("Age1", "Age2", "Age3")

expector(data, estimates)
}
