% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{initialise}
\alias{initialise}
\title{Initialise Data for EM Algorithm}
\usage{
initialise(known, unknown, sorted_data)
}
\arguments{
\item{known}{Data frame with columns for Length and Age, only for known Age}

\item{unknown}{Data frame with columns for Length and Age, only for NA Age}

\item{sorted_data}{Data frame with columns for Length and Age, with all observations with known Age at the top, and all unknown Age obervations at the bottom.}
}
\value{
A data frame containing initial estimates for mu, sigma, and lambda for Age groups 1, 2, and 3
}
\description{
Initialise Data for EM Algorithm
}
\examples{
known <- docExampleData[!is.na(docExampleData$Age), ] # known data is not NA
unknown <- docExampleData[is.na(docExampleData$Age), ] # unknown data is NA
sorted_data <- rbind(known, unknown) # combine data

initialise(known, unknown, sorted_data)
}
