% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{teamEM}
\alias{teamEM}
\title{Perform EM Optimisation}
\usage{
teamEM(data, epsilon = 1e-08, maxit = 1000)
}
\arguments{
\item{data}{Data frame with columns for Length and Age, including unknown values.}

\item{epsilon}{Convergence threshold for when the change in log likelihoods is < epsilon}

\item{maxit}{Maximum iterations allowed before algorithm halts optimisation}
}
\value{
A list of results from the optimisation:
\itemize{
 \item estimates - Data frame of final estimates for mu, sigma, and lambda for Age groups 1, 2, and 3
 \item inits - Data frame of initialisation values for mu, sigma, and lambda for Age groups 1, 2, and 3
 \item converged - TRUE/FALSE did the algorithm converge on values given the convergence threshold epsilon
 \item posteriors - Data frame of fnial posterior probabilities of each observed fish length belonging to its assigned Age group
 \item likelihood - Vector of likelihood values for each iteration in the algorithm
}
}
\description{
Perform EM Optimisation
}
\examples{
teamEM(docExampleData)
}
