\name{x}
\alias{FishLengths}
\docType{data}
\title{Fish Lengths Data}
\description{Data of fish of different lengths and their associated age group. 100 observations contain an age group assignments, 900 observations contain an unassigned age group.}
\usage{data("x")}
\format{
  A data frame with 1000 observations on the following 3 variables.
  \describe{
    \item{\code{FishID}}{a numeric vector}
    \item{\code{Length}}{a numeric vector}
    \item{\code{Age}}{a numeric vector}
  }
}
\source{https://moody.st-andrews.ac.uk/moodle/mod/resource/view.php?id=1267766}
\keyword{datasets}
